clipshare-backend/requirements.txt
Flask==3.0.0
flask-cors==4.0.0
azure-cosmos==4.5.1
azure-storage-blob==12.19.0
azure-cognitiveservices-vision-computervision==0.9.0
gunicorn==21.2.0
requests==2.31.0
python-dotenv==1.0.0

